+++ test/unit/replication-graphql.test.ts
@@ -818,6 +818,29 @@ describe('replication-graphql.test.js', () => {
                 server.close();
                 c.database.destroy();
             });
+            it('should pull documents from a custom dataPath if one is specified', async () => {
+                const [c, server] = await Promise.all([
+                    humansCollection.createHumanWithTimestamp(0),
+                    SpawnServer.spawn(getTestData(batchSize))
+                ]);
+                const replicationState = c.syncGraphQL({
+                    url: server.url,
+                    pull: {
+                        queryBuilder,
+                        dataPath: 'data.feedForRxDBReplication'
+                    },
+                    deletedFlag: 'deleted'
+                });
+                assert.strictEqual(replicationState.isStopped(), false);
+
+                await AsyncTestUtil.waitUntil(async () => {
+                    const docs = await c.find().exec();
+                    return docs.length === batchSize;
+                });
+
+                server.close();
+                c.database.destroy();
+            });
             it('pulled docs should be marked with a special revision if syncRevisions is false', async () => {
                 const [c, server] = await Promise.all([
                     humansCollection.createHumanWithTimestamp(0),